adds validateUser function in middleware\user.js with the password validation
